+++ src/test/java/org/fundacionjala/pivotal/steps/ui/story/StorySteps.java
@@ -1,33 +1,34 @@
 package org.fundacionjala.pivotal.steps.ui.story;
 
+import java.util.Map;
+
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.When;
+
 import org.fundacionjala.pivotal.pages.dashboard.DashBoard;
 import org.fundacionjala.pivotal.pages.story.DescriptionsStory;
 import org.fundacionjala.pivotal.pages.story.Story;
 import org.fundacionjala.pivotal.util.Helper;
 
-import java.util.Map;
-
 import static org.fundacionjala.pivotal.pages.story.DescriptionsStory.STORY_NAME;
 
 /**
- * CreateStorySteps class that create a story.
+ * StorySteps class that create a story.
  */
-public class CreateStorySteps {
-    private Story createStory;
+public class StorySteps {
+    private Story story;
     private Helper feature;
     private DashBoard dashBoard;
 
     /**
-     * Constructor for CreateStorySteps class.
+     * Constructor for StorySteps class.
      *
-     * @param createStory object.
+     * @param story object.
      * @param feature     object.
      * @param dashBoard   object.
      */
-    public CreateStorySteps(final Story createStory, final Helper feature, final DashBoard dashBoard) {
-        this.createStory = createStory;
+    public StorySteps(final Story story, final Helper feature, final DashBoard dashBoard) {
+        this.story = story;
         this.feature = feature;
         this.dashBoard = dashBoard;
     }
@@ -39,9 +40,9 @@ public class CreateStorySteps {
      */
     @When("^I create a new story$")
     public void iCreateANewStory(final Map<DescriptionsStory, String> values) {
-        createStory.setButtonCreateStory();
-        values.keySet().stream().forEach(step -> createStory.getStrategyStepMap(values).get(step).execute());
-        createStory.setButtonSaveStory();
+        story.setButtonCreateStory();
+        values.keySet().forEach(step -> story.getStrategyStepMap(values).get(step).execute());
+        story.setButtonSaveStory();
         feature.setStoryName(values.get(STORY_NAME));
     }
 
@@ -53,5 +54,28 @@ public class CreateStorySteps {
         dashBoard.setButtonDashborad();
         dashBoard.searchOnlyProject(feature.getProjectName());
         dashBoard.setProjectTitle();
+    }
+
+    /**
+     * First step to delete a story.
+     */
+    @Given("^I delete the story$")
+    public void iDeleteTheStory() {
+        story.deleteStory();
+    }
+
+
+
+    /**
+     * First step to edit a story.
+     *
+     * @param values parameters to edit.
+     */
+    @When("^I edit the next parameters$")
+    public void iEditTheNextParameters(final Map<DescriptionsStory, String> values) {
+        story.clickSToryExpander();
+        values.keySet().forEach(step -> story.getStrategyStepMap(values).get(step).execute());
+        story.setButtonCloseEdit();
+        feature.setStoryName(values.get(STORY_NAME));
     }
 }
